# OpenGL optimizations