--- a/internal/driver/mobile/gl/fn.go
+++ b/internal/driver/mobile/gl/fn.go
@@ -39,6 +39,7 @@ const (
 	glfnBlendColor
 	glfnBlendFunc
 	glfnBufferData
+	glfnBufferSubData
 	glfnClear
 	glfnClearColor
 	glfnCompileShader