--- a/internal/driver/mobile/gl/work.h
+++ b/internal/driver/mobile/gl/work.h
@@ -45,6 +45,7 @@ typedef enum {
 	glfnBlendColor,
 	glfnBlendFunc,
 	glfnBufferData,
+    glfnBufferSubData,
 	glfnClear,
 	glfnClearColor,
 	glfnCompileShader,