--- a/internal/painter/gl/gl_core.go
+++ b/internal/painter/gl/gl_core.go
@@ -72,10 +72,64 @@ func (p *painter) Init() {
 	gl.Disable(gl.DEPTH_TEST)
 	gl.Enable(gl.BLEND)
 	p.logError()
-	p.program = p.createProgram("simple")
-	p.lineProgram = p.createProgram("line")
-	p.rectangleProgram = p.createProgram("rectangle")
-	p.roundRectangleProgram = p.createProgram("round_rectangle")
+	p.program = ProgramState{
+		ref:        p.createProgram("simple"),
+		buff:       p.createBuffer(20),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(p.program, "text", "alpha")
+	p.enableAttribArrays(p.program, "vert", "vertTexCoord")
+
+	p.lineProgram = ProgramState{
+		ref:        p.createProgram("line"),
+		buff:       p.createBuffer(24),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(p.lineProgram, "feather", "color", "lineWidth")
+	p.enableAttribArrays(p.lineProgram, "vert", "normal")
+
+	p.rectangleProgram = ProgramState{
+		ref:        p.createProgram("rectangle"),
+		buff:       p.createBuffer(16),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(
+		p.rectangleProgram,
+		"frame_size", "rect_coords", "stroke_width", "fill_color", "stroke_color",
+	)
+	p.enableAttribArrays(p.rectangleProgram, "vert", "normal")
+
+	p.roundRectangleProgram = ProgramState{
+		ref:        p.createProgram("round_rectangle"),
+		buff:       p.createBuffer(16),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(p.roundRectangleProgram,
+		"frame_size", "rect_coords",
+		"stroke_width_half", "rect_size_half",
+		"radius", "edge_softness",
+		"fill_color", "stroke_color",
+	)
+	p.enableAttribArrays(p.roundRectangleProgram, "vert", "normal")
+}
+
+func (p *painter) getUniformLocations(pState ProgramState, names ...string) {
+	for _, name := range names {
+		u := p.ctx.GetUniformLocation(pState.ref, name)
+		pState.uniforms[name] = &UniformState{ref: u}
+	}
+}
+
+func (p *painter) enableAttribArrays(pState ProgramState, names ...string) {
+	for _, name := range names {
+		a := p.ctx.GetAttribLocation(pState.ref, name)
+		p.ctx.EnableVertexAttribArray(a)
+		pState.attributes[name] = a
+	}
 }
 
 type coreContext struct{}
@@ -108,6 +162,10 @@ func (c *coreContext) BlendFunc(srcFactor, destFactor uint32) {
 
 func (c *coreContext) BufferData(target uint32, points []float32, usage uint32) {
 	gl.BufferData(target, 4*len(points), gl.Ptr(points), usage)
+}
+
+func (c *coreContext) BufferSubData(target uint32, points []float32) {
+	gl.BufferSubData(target, 0, 4*len(points), gl.Ptr(points))
 }
 
 func (c *coreContext) Clear(mask uint32) {