--- a/internal/painter/gl/gl_wasm.go
+++ b/internal/painter/gl/gl_wasm.go
@@ -63,10 +63,64 @@ func (p *painter) Init() {
 	gl.Disable(gl.DEPTH_TEST)
 	gl.Enable(gl.BLEND)
 	p.logError()
-	p.program = p.createProgram("simple_es")
-	p.lineProgram = p.createProgram("line_es")
-	p.rectangleProgram = p.createProgram("rectangle_es")
-	p.roundRectangleProgram = p.createProgram("round_rectangle_es")
+	p.program = ProgramState{
+		ref:        p.createProgram("simple_es"),
+		buff:       p.createBuffer(20),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(p.program, "text", "alpha")
+	p.enableAttribArrays(p.program, "vert", "vertTexCoord")
+
+	p.lineProgram = ProgramState{
+		ref:        p.createProgram("line_es"),
+		buff:       p.createBuffer(24),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(p.lineProgram, "color", "feather", "lineWidth")
+	p.enableAttribArrays(p.lineProgram, "vert", "normal")
+
+	p.rectangleProgram = ProgramState{
+		ref:        p.createProgram("rectangle_es"),
+		buff:       p.createBuffer(16),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(
+		p.rectangleProgram,
+		"frame_size", "rect_coords", "stroke_width", "fill_color", "stroke_color",
+	)
+	p.enableAttribArrays(p.rectangleProgram, "vert", "normal")
+
+	p.roundRectangleProgram = ProgramState{
+		ref:        p.createProgram("round_rectangle_es"),
+		buff:       p.createBuffer(16),
+		uniforms:   make(map[string]*UniformState),
+		attributes: make(map[string]Attribute),
+	}
+	p.getUniformLocations(p.roundRectangleProgram,
+		"frame_size", "rect_coords",
+		"stroke_width_half", "rect_size_half",
+		"radius", "edge_softness",
+		"fill_color", "stroke_color",
+	)
+	p.enableAttribArrays(p.roundRectangleProgram, "vert", "normal")
+}
+
+func (p *painter) getUniformLocations(pState ProgramState, names ...string) {
+	for _, name := range names {
+		u := p.ctx.GetUniformLocation(pState.ref, name)
+		pState.uniforms[name] = &UniformState{ref: u}
+	}
+}
+
+func (p *painter) enableAttribArrays(pState ProgramState, names ...string) {
+	for _, name := range names {
+		a := p.ctx.GetAttribLocation(pState.ref, name)
+		p.ctx.EnableVertexAttribArray(a)
+		pState.attributes[name] = a
+	}
 }
 
 type xjsContext struct{}
@@ -99,6 +153,11 @@ func (c *xjsContext) BlendFunc(srcFactor, destFactor uint32) {
 
 func (c *xjsContext) BufferData(target uint32, points []float32, usage uint32) {
 	gl.BufferData(gl.Enum(target), toLEByteOrder(points...), gl.Enum(usage))
+}
+
+func (c *xjsContext) BufferSubData(target uint32, points []float32) {
+	data := toLEByteOrder(points...)
+	gl.BufferSubData(gl.Enum(target), 0, data)
 }
 
 func (c *xjsContext) Clear(mask uint32) {