<!-- If this is your first pull request for Fyne please read the contributor docs at:
https://github.com/fyne-io/fyne/wiki/Contributing.
Be sure that your work is based off `develop` branch. --> 

### Description:
As an attempt at fixing #1062, this PR significantly reduces the number of OpenGL calls needed during a paint event.
I have tested on android, Linux desktop and WASM, but it still needs testing on IOS, macOS and Windows. Tests are passing on all platforms, however. 

Testing using the example code from #4058.

#### Optimizations
1. Cache uniform locations during initialization, so they don't need to be looked up during every draw.
2. Only set uniforms if the uniform data has changed.
3. Only create buffer objects during initialization, after that use `BufferSubData` to update the buffer.
4. Enable attributes during initialization, and only update them during rendering. 

On a MOTO G 2014:
Without optimizations (develop branch): ~150ms and ~2970 OpenGL calls 
With optimizations: ~50ms and ~1240 OpenGL calls

I also did a quick test on Samsung S24, and paints went from ~50ms to ~18ms.

I also tested https://github.com/fyne-io/demo as well.

### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [ ] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.