@andydotxyz 2025-08-07
Tested on iOS and Darwin. Everything is working correctly. 
However on iOS the scrolling seems to have got "stuttery" - see if you can catch what I mean in these videos.
Performance overall is better but this feels less smooth now for some reason.

develop:

https://github.com/user-attachments/assets/3df98936-4565-48f8-8ae4-39d9c40dc446

PR 5884:

https://github.com/user-attachments/assets/ef484332-53b3-4240-8ac4-78709e563724