@andydotxyz 2025-08-09
> Is it possible this effect is caused by decreased paint time during scrolling? I noticed during testing that the current mobile scroll decay logic is sensitive to how long the paint takes, because it waits for the paint to finish before sleeping for 16ms until the next Dragged event is dispatched:

Good catch!

I tested with this patch applied and it is very smooth, looks like an OK adjustment now that drawing is faster :)

```diff
index 47564ac96..b8ebc4f2b 100644
--- a/internal/driver/mobile/driver.go
+++ b/internal/driver/mobile/driver.go
@@ -466,11 +466,11 @@ func (d *driver) tapUpCanvas(w *window, x, y float32, tapID touch.Sequence) {
 
                                d.DoFromGoroutine(func() {
                                        wid.Dragged(ev)
-                               }, true)
+                               }, false)
                                time.Sleep(time.Millisecond * 16)
                        }
 
-                       d.DoFromGoroutine(wid.DragEnd, true)
+                       d.DoFromGoroutine(wid.DragEnd, false)
                }()
        })
 }
```