@andydotxyz 2025-08-09
Windows looks good too