@redawl 2025-08-10
> > Is it possible this effect is caused by decreased paint time during scrolling? I noticed during testing that the current mobile scroll decay logic is sensitive to how long the paint takes, because it waits for the paint to finish before sleeping for 16ms until the next Dragged event is dispatched:
> 
> Good catch!
> 
> I tested with this patch applied and it is very smooth, looks like an OK adjustment now that drawing is faster :)
> 
Nice! Want me to include that patch in this PR?