@andydotxyz 2025-08-10
> Nice! Want me to include that patch in this PR?

Yes please otherwise iOS feels like degraded performance.